Move libr/include/r_rbtree.h to libr/include/r_util/r_rbtree.h
