+++ libr/util/rbtree.c
@@ -1,4 +1,4 @@
-#include "r_rbtree.h"
+#include "r_util/r_rbtree.h"
 
 static void rbnode_clear(RListFree freefn, RBNode *x) {
 	if (x) {
@@ -331,88 +331,3 @@ R_API void *r_rbtree_iter_next(RBTreeIter *it) {
 R_API void *r_rbtree_iter_prev(RBTreeIter *it) {
 	return next (it, 1);
 }
-
-#if TEST
-// TODO: move into t/rbtree.c
-#include <assert.h>
-#include <stddef.h>
-#include <stdlib.h>
-
-void random_iota(int *a, int n) {
-  a[0] = 0;
-  for (int i = 1; i < n; i++) {
-    int x = rand () % (i + 1);
-    if (i != x) {
-      a[i] = a[x];
-    }
-    a[x] = i;
-  }
-}
-
-int cmp_int(const void *a, const void *b) {
-  return (int)(ptrdiff_t)a - (int)(ptrdiff_t)b;
-}
-
-int main() {
-#define N 1000
-	int a[N], i;
-	random_iota (a, N);
-
-	RBTree *tree = r_rbtree_new (NULL, cmp_int);
-	RBTreeIter it;
-	void *data;
-
-	// Insert
-	for (int i = 0; i < N; i++) {
-		assert (r_rbtree_insert (tree, (void *)(ptrdiff_t)a[i]) == 1);
-		assert (r_rbtree_insert (tree, (void *)(ptrdiff_t)a[i]) == 0); // returns 0 because it exists
-		assert (tree->size == i + 1);
-	}
-
-	//r_rbtree_print (tree);
-
-	const int key = 0x24;
-	// lower_bound
-	it = r_rbtree_lower_bound_forward (tree, (void *)(ptrdiff_t)key);
-	i = key;
-	r_rbtree_iter_while(it, data) {
-		assert ((ptrdiff_t)data == i++);
-	}
-	it = r_rbtree_lower_bound_backward (tree, (void *)(ptrdiff_t)key);
-	i = key - 1;
-	r_rbtree_iter_while_prev(it, data) {
-		assert ((ptrdiff_t)data == i--);
-	}
-
-	// upper_bound
-	it = r_rbtree_upper_bound_forward (tree, (void *)(ptrdiff_t)key);
-	i = key + 1;
-	r_rbtree_iter_while(it, data) {
-		assert ((ptrdiff_t)data == i++);
-	}
-	it = r_rbtree_upper_bound_backward (tree, (void *)(ptrdiff_t)key);
-	i = key;
-	r_rbtree_iter_while_prev(it, data) {
-		assert ((ptrdiff_t)data == i--);
-	}
-
-	// Traverse
-	i = 0;
-	r_rbtree_foreach (tree, it, data) {
-		assert ((ptrdiff_t)data == i++);
-	}
-	r_rbtree_foreach_prev (tree, it, data) {
-		assert ((ptrdiff_t)data == --i);
-	}
-
-	// Delete
-	random_iota (a, N);
-	for (int i = 0; i < N; i++) {
-		assert (r_rbtree_delete (tree, (void *)(ptrdiff_t)a[i]));
-		assert (!r_rbtree_delete (tree, (void *)(ptrdiff_t)a[i]));
-		assert (tree->size == N - i - 1);
-	}
-
-	r_rbtree_free (tree);
-}
-#endif
